+++ test/clj_kondo/main_test.clj
@@ -3132,7 +3132,7 @@ foo/")))
 
 (deftest as-alias-test
   (assert-submaps
-   '({:file "<stdin>", :row 1, :col 20, :level :warning, :message "namespace foo.bar is required but never used"}
+   '(#_{:file "<stdin>", :row 1, :col 20, :level :warning, :message "namespace foo.bar is required but never used"}
      {:file "<stdin>", :row 1, :col 44, :level :warning, :message "Unresolved namespace fx. Are you missing a require?"})
    (lint! "(ns foo (:require [foo.bar :as-alias fb])) ::fx/bar"))
   (is (empty? (lint! "(ns foo (:require [foo.bar :as-alias fb])) ::fb/bar"))))