+++ test/clj_kondo/test_test.clj
@@ -6,23 +6,25 @@
    [clojure.test :refer [deftest is]]))
 
 (deftest missing-test-assertion-test
-  (is (empty? (lint! "(ns foo (:require [clojure.test :as t])) (t/deftest (t/is (odd? 1)))")))
-  (is (empty? (lint! "(ns foo (:require [clojure.test :as t])) (t/deftest- (t/is (odd? 1)))")))
-  (assert-submaps
-   '({:file "<stdin>", :row 1, :col 57, :level :warning, :message "missing test assertion"})
-   (lint! "(ns foo (:require [clojure.test :as t])) (t/deftest foo (odd? 1))"))
-  (assert-submaps
-   '({:file "<stdin>", :row 2, :col 21, :level :warning, :message "missing test assertion"})
-   (lint! "(ns foo (:require [clojure.test :as t] [clojure.set :as set]))
+  (doseq [lang ["clj" "cljs"]]
+    (let [lint! #(lint! % "--lang" lang )]
+      (is (empty? (lint! "(ns foo (:require [clojure.test :as t])) (t/deftest (t/is (odd? 1)))")))
+      (is (empty? (lint! "(ns foo (:require [clojure.test :as t])) (t/deftest- (t/is (odd? 1)))")))
+      (assert-submaps
+       '({:file "<stdin>", :row 1, :col 57, :level :warning, :message "missing test assertion"})
+       (lint! "(ns foo (:require [clojure.test :as t])) (t/deftest foo (odd? 1))"))
+      (assert-submaps
+       '({:file "<stdin>", :row 2, :col 21, :level :warning, :message "missing test assertion"})
+       (lint! "(ns foo (:require [clojure.test :as t] [clojure.set :as set]))
      (t/deftest foo (set/subset? #{1 2} #{1 2 3}))"))
-  (assert-submaps
-   '({:file "<stdin>", :row 2, :col 38, :level :warning, :message "missing test assertion"})
-   (lint! "(ns foo (:require [clojure.test :as t] [clojure.set :as set]))
+      (assert-submaps
+       '({:file "<stdin>", :row 2, :col 38, :level :warning, :message "missing test assertion"})
+       (lint! "(ns foo (:require [clojure.test :as t] [clojure.set :as set]))
      (t/deftest foo (t/testing \"foo\" (set/subset? #{1 2} #{1 2 3})))"))
-  (assert-submaps
-   '({:file "<stdin>", :row 2, :col 49, :level :warning, :message "missing test assertion"})
-   (lint! "(ns foo (:require [clojure.test :as t] [clojure.set :as set]))
-     (t/deftest foo (t/testing \"foo\" (let [x 1] (set/subset? #{1 2} #{1 2 3}))))")))
+      (assert-submaps
+       '({:file "<stdin>", :row 2, :col 49, :level :warning, :message "missing test assertion"})
+       (lint! "(ns foo (:require [clojure.test :as t] [clojure.set :as set]))
+     (t/deftest foo (t/testing \"foo\" (let [x 1] (set/subset? #{1 2} #{1 2 3}))))")))))
 
 (deftest redefined-test-test
   (assert-submaps